+++ ft_utility_base.c
@@ -6,13 +6,13 @@
 /*   By: csantivi <csantivi@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/11 12:55:44 by csantivi          #+#    #+#             */
-/*   Updated: 2022/03/11 13:31:56 by csantivi         ###   ########.fr       */
+/*   Updated: 2022/03/11 22:39:44 by csantivi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_print_p(unsigned long long p)
+int	ft_print_p(unsigned long p)
 {
 	int		size;
 	char	*str;
